+++ RR-Calibration/src/main/java/com/scor/rr/rest/adjustment/AdjustmentThreadRest.java
@@ -27,7 +27,7 @@ public class AdjustmentThreadRest {
     }
 
     @GetMapping
-    public AdjustmentThreadEntity findById(Long id){
+    public AdjustmentThreadEntity findById(Integer id){
         return adjustmentThreadService.findOne(id);
     }
