+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentEventBasedParameterRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface AdjustmentEventBasedParameterRepository  extends JpaRepository<AdjustmentEventBasedParameterEntity, Long> {
+public interface AdjustmentEventBasedParameterRepository  extends JpaRepository<AdjustmentEventBasedParameterEntity, Integer> {
 //    @Query("select p from AdjustmentEventBasedParameterEntity p inner join AdjustmentNodeEntity n where p.adjustmentNodeByFkAdjustmentNodeId = n and n.adjustmentNodeId = :id")
 //    AdjustmentEventBasedParameterEntity findByAdjustmentNodeAdjustmentNodeId(@Param("id") int idAdjustmentNode);
     AdjustmentEventBasedParameterEntity findByAdjustmentNodeAdjustmentNodeId(Long adjustmentNodeId);
 
-    void deleteByAdjustmentNodeAdjustmentNodeId(Long adjustmentNodeId);
+    void deleteByAdjustmentNodeAdjustmentNodeId(Integer adjustmentNodeId);
 }
