+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentStateRepository.java
@@ -4,7 +4,7 @@ import com.scor.rr.domain.AdjustmentState;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AdjustmentStateRepository extends JpaRepository<AdjustmentState, Long>{
+public interface AdjustmentStateRepository extends JpaRepository<AdjustmentState, Integer>{
     @Query("select m from AdjustmentState m where m.code = 'Invalid'")
     AdjustmentState getAdjustmentStateEntityByCodeInvalid();
 }
