+++ RR-Common/src/main/java/com/scor/rr/repository/AdjustmentTypeRepository.java
@@ -3,5 +3,5 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.AdjustmentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdjustmentTypeRepository extends JpaRepository<AdjustmentType, Long> {
+public interface AdjustmentTypeRepository extends JpaRepository<AdjustmentType, Integer> {
 }
