+++ RR-Common/src/main/java/com/scor/rr/repository/PltHeaderRepository.java
@@ -3,7 +3,7 @@ package com.scor.rr.repository;
 import com.scor.rr.domain.PltHeaderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PltHeaderRepository extends JpaRepository<PltHeaderEntity, Long> {
-    PltHeaderEntity findByPltHeaderId(Long pkScorPltHeaderId);
+public interface PltHeaderRepository extends JpaRepository<PltHeaderEntity, Integer> {
+    PltHeaderEntity findByPltHeaderId(Integer pkScorPltHeaderId);
 }
